+++ EventManager.Server/Interfaces/IEventRepository.cs
@@ -5,7 +5,7 @@ namespace EventManager.Server.Interfaces
 {
     public interface IEventRepository
     {
-        public Task<Event?> GetEventByIdAsync(ulong id);
+        public Task<Event?> GetEventByIdAsync(int id);
         public Task<PagedResult<Event>> GetEventsAsync(PaginationQuery query);
         public Task<Event> CreateEventAsync(Event newEvent);
         public Task<Event> UpdateEventAsync(Event updatedEvent);
