+++ EventManager.Server/Interfaces/IEventService.cs
@@ -5,10 +5,10 @@ namespace EventManager.Server.Interfaces
 {
     public interface IEventService
     {
-        public Task<EventDto?> GetEventByIdAsync(ulong id);
+        public Task<EventDto?> GetEventByIdAsync(int id);
         public Task<PagedResult<EventDto>> GetEventsAsync(PaginationQuery query);
         public Task<EventDto> CreateEventAsync(EventCreateDto dto);
         public Task<EventDto?> UpdateEventAsync(EventDto dto);
-        public Task<bool> DeleteEventAsync(ulong id);
+        public Task<bool> DeleteEventAsync(int id);
     }
 }
